challenge_1/pyproject.toml
[tool.poetry]
name = "challenge-1"
version = "1.0.0"
description = ""
authors = ["Filippo Manzardo <film@bendingspoons.com>", "Leonardo La Rocca <leol@bendingspoons.com>", "Igor Katselenbogen <ik@bendingspoons.com>"]
readme = "README.md"
packages = [{include = "challenge_1"}]


[tool.poetry.scripts]
train = "challenge_1.main:run_training"

[tool.poetry.dependencies]
python = "~3.10"

google-cloud = "^0.34.0"
google-cloud-storage = "^2.6.0"
keras-tuner = "^1.1.3"
matplotlib = "^3.2.2"
numpy = "^1.21.6"
pandas = "^1.3.5"
tensorflow = {version = "^2.10.0", markers = "sys_platform != 'darwin'"}
tensorflow-macos = {version = "^2.10.0", markers = "sys_platform == 'darwin'"}
scipy = "^1.7.3"
scikit-image = "^0.18.3"
scikit-learn = "^1.0.2"
seaborn = "^0.11.2"
split-folders = "^0.5.1"
tqdm = "^4.64.1"


[tool.poetry.group.dev.dependencies]
colorlog = "^6.7.0"
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
mypy = "^0.990"
pytest = "^7.1.1"
click = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
exclude = '''
(
    \.idea |
    \.DS_Store |
    \.githooks |
    \.vscode |
    venv |
    .html |
    .json |
    .pyc |
    .sql
)
'''

[tool.isort]
profile = "black"
line_length = 100
